Separate out the CLI into core, interactive, and non-interactive. - Start allowing some commands to be issued non-interactively, from the compiled command line interface. - Some commands will not be allowed to be used interactively at this time due to complexity / additional refactoring required to use them.
