+++ src/build_terminal.js
@@ -7,12 +7,12 @@ import {JsonRpcServer} from './node/jsonRpc';
 
 JsonRpcServer(consts.JSON_RPC);
 
-process.on('uncaughtException', function (err) {
-    console.error('Caught exception: ' + err);
+process.on('uncaughtException', (err) => {
+    console.error(`Caught exception: ${err}`);
 });
 
 //                            light-node
-Blockchain.createBlockchain('full-node', ()=>{}, async ()=>{
+Blockchain.createBlockchain('full-node', () => {}, async () => {
 
     await Node.NodeExpress.startExpress();
 
@@ -22,14 +22,10 @@ Blockchain.createBlockchain('full-node', ()=>{}, async ()=>{
     Node.NodeClientsService.startService();
 
     Node.NodeServer.startServer();
-
-
-}, ()=>{
+}, () => {
 });
 
 
 process.on('SIGINT', async () => {
-
     await termination(Blockchain);
-
 });
