+++ src/node/menu/CLI-Menu.js
@@ -1,10 +1,4 @@
 /* eslint-disable */
-import CONNECTION_TYPE from "../lists/types/Connection-Type";
-import NODES_CONSENSUS_TYPE from "../lists/types/Node-Consensus-Type";
-
-
-const FileSystem = require('fs');
-const readline = require('readline');
 import {JsonRpcServer} from './../jsonRpc';
 
 let NodeExpress, NodeServer;
@@ -16,17 +10,18 @@ if (!process.env.BROWSER) {
 
 
 import consts from 'consts/const_global';
-import {Node} from '../../index.js';
 import AdvancedMessages from './Advanced-Messages';
-import WebDollarCoins from "common/utils/coins/WebDollar-Coins";
-import InterfaceBlockchainAddressHelper from "common/blockchain/interface-blockchain/addresses/Interface-Blockchain-Address-Helper";
 import Blockchain from "main-blockchain/Blockchain"
-import StatusEvents from "common/events/Status-Events";
 import NodeServer from 'node/sockets/node-server/sockets/Node-Server';
 import Log from 'common/utils/logging/Log';
-import AddressBanList from "common/utils/bans/AddressBanList";
+import CLICore from "./CLI-Core";
+import CLIRunner from './CLI-Menu-non-interactive';
 
 class CLI {
+
+    static get CORE() {
+        return CLICore.INSTANCE;
+    }
 
     constructor() {
 
@@ -35,7 +30,22 @@ class CLI {
 
         this._exitMenu = undefined;
 
-        this._start();
+        // parse cli args
+        this.cliArgs = process.argv;
+
+        if (!this.cliArgs.includes('--')) {
+            this._startInteractive();
+        } else {
+            console.log("Non-interactive activated.  Args will be read from left to right.");
+            this.cliArgs = this.cliArgs.splice(this.cliArgs.indexOf('--'));
+            if (this.cliArgs.length <= 0) {
+                CLI.CORE.showCommands();
+            }
+            console.log('Working with args: %o', this.cliArgs);
+            this.runner = new CLIRunner(this.cliArgs);
+            this.runner.run()
+                .then(() => process.exit());
+        }
 
     }
 
@@ -51,166 +61,90 @@ class CLI {
 
         switch (answer.trim()) {
             case '1': //  List addresses'
-                await this.listAddresses();
+                await CLI.CORE.listAddresses();
                 break;
             case '2': //  Create new address',
-                await this.createNewAddress();
+                await CLI.CORE.createNewAddress();
                 break;
             case '3': //  Delete Address
-                await this.deleteAddress();
+                await CLI.CORE.deleteAddress();
                 break;
             case '4': //  Import Address
-                await this.importAddress();
+                await CLI.CORE.importAddress();
                 break;
             case '5': //  Export Address
-                await this.exportAddress();
+                await CLI.CORE.exportAddress();
                 break;
             case '6': //  Encrypt Address
-                await this.encryptAddress();
+                await CLI.CORE.encryptAddress();
                 break;
             case '7': //  Set Mining Address
-                await this.setMiningAddress();
+                await CLI.CORE.setMiningAddress();
                 break;
             case '8': //  Start Mining
-                await this.startMining();
+                await CLI.CORE.startMining();
                 break;
             case '9': //  Start Mining Instantly
-                await this.startMining(true);
+                await CLI.CORE.startMining(true);
                 break;
             case '10': // Mining Pool: Start Mining in a Pool
-                await this.startMiningInsidePool();
+                await CLI.CORE.startMiningInsidePool();
                 break;
             case '11-1':  // Mining Pool: Create a New Pool
-                await this.processRemainingPayment();
+                await CLI.CORE.processRemainingPayment();
                 break;
             case '11':  // Mining Pool: Create a New Pool
-                await this.createMiningPool();
+                await CLI.CORE.createMiningPool();
                 break;
             case '12':  // Server Mining Pool: Create a new Server for Mining Pool
-                await this.createServerForMiningPool();
+                await CLI.CORE.createServerForMiningPool();
                 break;
             case '13': //  Import Address
-                await this.signTransaction();
+                await CLI.CORE.signTransaction();
                 break;
             case '20':  // Server Mining Pool: Create a new Server for Mining Pool
                 await NodeServer.startServer();
                 break;
             case '21': // Disable Forks Immutability
-                await this.disableForksImmutability();
+                await CLI.CORE.disableForksImmutability();
                 break;
             case '22': // Disable Forks Immutability
-                await this.disconnectFromAllConnectedNodes();
+                await CLI.CORE.disconnectFromAllConnectedNodes();
                 break;
             case '23':
-                await this.disconnectAllMinersNodes();
+                await CLI.CORE.disconnectAllMinersNodes();
                 break;
             case '24':
-                await this.setIntervalDisconnectAllMinersNodes();
+                await CLI.CORE.setIntervalDisconnectAllMinersNodes();
                 break;
             case '30':  // Set Password
                 await Blockchain.Mining.setPrivateKeyAddressForMiningAddress();
                 break;
             case '53': // add banlist
-                await this.AddAddressBanList();
+                await CLI.CORE.AddAddressBanList();
                 break;
             case '54': // add banlist
-                await this.LoadAddressBanList();
+                await CLI.CORE.LoadAddressBanList();
                 break;
             case '55': // add banlist
-                await this.SaveAddressBanList();
+                await CLI.CORE.SaveAddressBanList();
                 break;
             case 'exit':
                 this._exitMenu = true;
                 break;
             default:
-                this._showCommands();
+                CLI.CORE.showCommands();
                 break;
         }
 
         await this._runMenu();
     };
 
-    signTransaction() {
-
-        console.info('Sign Transaction');
-
-        return new Promise(async (resolve) => {
-
-            let addressId = await this._chooseAddress();
-
-            if (addressId < 0) {
-                console.warn("You must enter a valid number.");
-                resolve(false);
-                return;
-            }
-
-            let toAddress = await AdvancedMessages.input('Enter the recipient address: ');
-            let amountToSend = await AdvancedMessages.input('Enter the transaction amount: ');
-            let nonce = await AdvancedMessages.input('Enter the address current nonce: ');
-            let timelock = await AdvancedMessages.input('Enter the current block: ');
-            let addressPath = await AdvancedMessages.input('Enter path for saving the transaction:');
-            let wantToPropagate = await AdvancedMessages.input('Do you want to propagate now y/n?:');
-
-            if (wantToPropagate.toUpperCase().trim() === 'Y' ? true : false)
-                wantToPropagate = true;
-            else
-                wantToPropagate = false;
-
-            let feeToSend = Blockchain.Transactions.wizard.calculateFeeSimple(amountToSend);
-
-            let addressString = Blockchain.Wallet.addresses[addressId].address;
-            let answer = null;
-
-            //Trick for blocks length and address nonce
-            Blockchain.blockchain.blocks._length = timelock + 1;
-
-            for (let i = 0; i < Blockchain.Wallet.addresses.length; i++)
-                if (addressString === Blockchain.Wallet.addresses[i].address)
-                    answer = await Blockchain.Transactions.wizard.validateTransaction(Blockchain.Wallet.addresses[i].address, toAddress, amountToSend * WebDollarCoins.WEBD, feeToSend, undefined, undefined, timelock - 1, nonce, true);
-
-            let data = {};
-
-            if (answer.result) {
-
-                data.transaction = answer.transaction.serializeTransaction();
-                data.signature = answer.signature;
-
-                if (wantToPropagate)
-                    await Blockchain.blockchain.transactions.pendingQueue.includePendingTransaction(answer.transaction, undefined, true);
-
-            } else {
-
-                console.log("Transaction was not created. " + answer.message);
-                resolve(false);
-                return;
-
-            }
-
-            FileSystem.writeFile(addressPath + "transaction.tx", JSON.stringify(data), 'utf8', (err) => {
-
-                if (err) {
-                    console.error(err);
-                    resolve(false);
-                    return;
-                }
-
-                console.log("Transaction successfully exported to ," + addressPath + "transaction.tx");
-
-                resolve(true);
-
-            });
-
-            resolve(true);
-
-        });
-
-    }
-
-    async _start() {
+    async _startInteractive() {
 
         Log.info('CLI menu started', Log.LOG_TYPE.CLI_MENU);
 
-        if (Blockchain )
+        if (Blockchain)
             await Blockchain.loadWallet();
 
         this._showCommands();
@@ -222,587 +156,7 @@ class CLI {
         await this._runMenu();
     }
 
-    async processRemainingPayment() {
-
-        await this._callCallbackBlockchainSync(undefined, undefined, undefined, async () => {
-            await Blockchain.PoolManagement.poolRemainingRewards.doPayout();
-        }, true);
-
-    }
-
-    async _chooseAddress() {
-
-        await this.listAddresses();
-
-        let addressId = await AdvancedMessages.readNumber('Choose the address number: ');
-
-        if (isNaN(addressId) || addressId < 0 || Blockchain.Wallet.addresses.length <= addressId)
-            return -1;
-
-        return addressId;
-    }
-
-    async AddAddressBanList() {
-
-        let addressWIF = await AdvancedMessages.input('Please input miner address: ');
-        if (addressWIF.length == 40) {
-            let unencodedAddress = InterfaceBlockchainAddressHelper.getUnencodedAddressFromWIF(addressWIF);
-            if (unencodedAddress) {
-                let duration = await AdvancedMessages.readNumber('Ban duration (hours): ');
-                let reason = await AdvancedMessages.input('Please input ban reason: ');
-                reason = reason.replace(';', '.,');
-                AddressBanList.addBan(unencodedAddress, duration * 3600 * 1000, reason);
-                AddressBanList.listBans();
-            }
-        }
-
-    }
-
-    async LoadAddressBanList() {
-
-        try {
-
-            const file = FileSystem.createReadStream('address-ban-list.txt');
-
-            const rl = readline.createInterface({
-                input: file,
-                crlfDelay: Infinity
-            });
-
-            rl.on('error',  (err) => {
-                console.error('Address ban list was not found');
-            });
-
-            file.on('error',  (err) => {
-                console.error('Address ban list was not found');
-            });
-
-            rl.on('line', (line) => {
-                let triplet = line.split(';');
-                if (triplet.length === 3) {
-
-                    let duration = parseFloat(triplet[0]);
-                    let addressWIF = triplet[1];
-                    let reason = triplet[2];
-
-                    if (addressWIF && addressWIF.length == 40) {
-                        let unencodedAddress = InterfaceBlockchainAddressHelper.getUnencodedAddressFromWIF(addressWIF);
-                        if (unencodedAddress)
-                            AddressBanList.addBan(unencodedAddress, duration * 3600 * 1000, reason);
-                    }
-                } else console.log("Invalid line in address-ban-list.txt", line);
-            });
-
-            rl.on('close', () => {
-                AddressBanList.listBans();
-            });
-
-        } catch (exception) {
-            console.error("Error loading banlist: ", exception);
-        }
-    }
-
-    async SaveAddressBanList() {
-        AddressBanList.saveBans();
-    }
-
-    _showCommands() {
-
-        console.info('\nChoose one of the following commands:');
-
-        for (let i = 0; i < commands.length; ++i) {
-            console.info(commands[i]);
-        }
-        console.log();
-
-        return true;
-    }
-
-    async listAddresses() {
-
-
-        console.warn("INFO: YOU NEED TO BE SYNC TO SEE THE VALUE OF YOUR WALLET!!!");
-        console.info('\nWallet addresses:');
-
-        let miningAddress = Blockchain.blockchain.mining.minerAddress;
-        if ( !miningAddress )
-            miningAddress = 'not specified';
-
-        console.log(addressHeader);
-        for (let i = 0; i < Blockchain.Wallet.addresses.length; ++i) {
-
-            let address = Blockchain.Wallet.addresses[i].address;
-
-            let balance = Blockchain.blockchain.accountantTree.getBalance(address, undefined);
-
-            balance = (balance === null) ? 0 : (balance / WebDollarCoins.WEBD);
-
-            if (address === miningAddress)
-                console.log(((i < 10) ? "|  *" : "| *") + i + "   |  " + address + "  | " + balance + lineSeparator);
-            else
-                console.log(((i < 10) ? "|   " : "|  ") + i + "   |  " + address + "  | " + balance + lineSeparator);
-
-        }
-
-        let balance = 0;
-        if (miningAddress !== 'not specified') {
-            balance = Blockchain.blockchain.accountantTree.getBalance(miningAddress, undefined);
-            balance = (balance === null) ? 0 : balance;
-
-            if (Blockchain.MinerPoolManagement && Blockchain.MinerPoolManagement.minerPoolStarted)
-                balance += Blockchain.MinerPoolManagement.minerPoolReward.total;
-
-            balance /= WebDollarCoins.WEBD;
-
-
-        }
-        console.log("| MINING|  " + miningAddress + "  | " + balance + lineSeparator);
-
-        return true;
-
-    }
-
-    async createNewAddress() {
-
-        console.info('Create new address.');
-        try {
-            let address = await Blockchain.Wallet.createNewAddress();
-            console.info("Address was created: " + address.address);
-            return true;
-        } catch (err) {
-            console.err(err);
-            return false;
-        }
-
-    }
-
-    async deleteAddress() {
-
-        console.info('Delete address.');
-
-        let addressId = await this._chooseAddress();
-
-        if (addressId < 0) {
-            console.warn("You must enter a valid number.");
-            return false;
-        }
-
-        let response = await Blockchain.Wallet.deleteAddress(Blockchain.Wallet.addresses[addressId].address);
-
-        return response.result;
-    }
-
-    importAddress() {
-
-        console.info('Import address.');
-
-        return new Promise(async (resolve) => {
-
-            let addressPath = await AdvancedMessages.input('Enter address path: ');
-
-            FileSystem.readFile(addressPath, 'utf8', async (err, content) => {
-
-                if (err) {
-                    console.error(err);
-                    resolve(false);
-                    return;
-                }
-
-                try {
-
-                    let answer = await Blockchain.Wallet.importAddressFromJSON(JSON.parse(content));
-
-                    if (answer.result === true) {
-                        console.log("Address successfully imported", answer.address);
-                        await Blockchain.Wallet.saveWallet();
-
-                        if (Blockchain.Wallet.addresses.length === 1) Blockchain.blockchain.mining.minerAddress = Blockchain.Wallet.addresses[0].address;
-
-                        resolve(true);
-                    } else {
-                        console.error(answer.message);
-                        resolve(false);
-                    }
-
-                } catch (err) {
-                    console.error(err.message);
-                    resolve(false);
-                }
-
-
-            });
-
-        });
-
-    }
-
-    exportAddress() {
-
-        console.info('Export address.');
-
-        return new Promise(async (resolve) => {
-
-            let addressId = await this._chooseAddress();
-
-            if (addressId < 0) {
-                console.warn("You must enter a valid number.");
-                resolve(false);
-                return;
-            }
-
-            let addressPath = await AdvancedMessages.input('Enter path for saving address: ');
-
-            let addressString = Blockchain.Wallet.addresses[addressId].address;
-            let fileName = "WEBD$" + Blockchain.Wallet.addresses[addressId].unencodedAddress.toString("hex") + ".webd";
-
-            let answer = await Blockchain.Wallet.exportAddressToJSON(addressString);
-
-            if (answer.result === false) {
-                console.log("Address was not exported. :(. " + answer.message);
-                resolve(false);
-                return;
-            }
-
-            let jsonAddress = JSON.stringify(answer.data);
-
-            FileSystem.writeFile(addressPath + fileName, jsonAddress, 'utf8', (err) => {
-
-                if (err) {
-                    console.error(err);
-                    resolve(false);
-                    return;
-                }
-
-                console.log("Address successfully exported", addressString, '   to ', addressPath + fileName);
-
-                resolve(true);
-                return;
-
-            });
-
-            resolve(true);
-            return;
-
-        });
-
-    }
-
-    async encryptAddress() {
-
-        console.info('Encrypt address.');
-
-        let addressId = await this._chooseAddress();
-
-        if (addressId < 0) {
-            console.warn("You must enter a valid number.");
-            return false;
-        }
-
-        let addressString = Blockchain.Wallet.addresses[addressId].address;
-        let newPassword = await InterfaceBlockchainAddressHelper.askForPassword("Please enter a password(12 words separated by space):");
-        let response = await Blockchain.Wallet.encryptAddress(addressString, newPassword);
-
-        if (response === true)
-            console.info("Address was encrypted:", addressString);
-        else
-            console.error("Address couldn't be encrypted:", addressString);
-
-        return response.result;
-    }
-
-    async setMiningAddress() {
-
-        console.info('Set mining address.');
-
-        let addressId = await this._chooseAddress();
-
-        if (addressId < 0) {
-            console.warn("You must enter a valid number.");
-            return false;
-        }
-
-        Blockchain.blockchain.mining.minerAddress = Blockchain.Wallet.addresses[addressId].address;
-
-        return true;
-    }
-
-    async startMining(instantly) {
-
-
-        await this._callCallbackBlockchainSync(undefined, () => Blockchain.MinerPoolManagement.minerPoolSettings.setMinerPoolActivated(false), async () => {
-
-            if (instantly)
-                await Blockchain.startMiningInstantly();
-            else
-                Blockchain.startMiningNextTimeSynchronized = true;
-
-        }, undefined, undefined);
-
-    }
-
-    async startMiningInsidePool() {
-
-        Log.info('Mining inside a POOL', Log.LOG_TYPE.POOLS);
-
-        consts.SETTINGS.NODE.PORT = consts.SETTINGS.NODE.MINER_POOL_PORT;
-
-        await this._callCallbackBlockchainSync(undefined, async () => {
-
-            try {
-
-                let getNewLink = true;
-
-                if (typeof Blockchain.MinerPoolManagement.minerPoolSettings.poolURL === "string" && Blockchain.MinerPoolManagement.minerPoolSettings.poolURL !== '') {
-
-                    Log.info('Your current mining pool is: ' + Blockchain.MinerPoolManagement.minerPoolSettings.poolName + " " + Blockchain.MinerPoolManagement.minerPoolSettings.poolWebsite, Log.LOG_TYPE.error);
-                    let response = await AdvancedMessages.confirm('Do you want to continue mining in the same pool: ' + Blockchain.MinerPoolManagement.minerPoolSettings.poolURL);
-
-                    if (response === true) getNewLink = false;
-                    7
-                }
-
-                let miningPoolLink = undefined;
-
-                if (getNewLink) {
-
-                    miningPoolLink = await AdvancedMessages.input('Enter the new mining pool link: ');
-                    Log.info('Your new MiningPool is : ' + miningPoolLink, Log.LOG_TYPE.info);
-
-                }
-
-                StatusEvents.on("miner-pool/connection-established", (data) => {
-                    if (data.connected)
-                        Blockchain.Mining.startMining();
-                    else
-                        Blockchain.Mining.stopMining();
-                });
-
-                Blockchain.MinerPoolManagement.startMinerPool(miningPoolLink, true);
-
-            } catch (exception) {
-
-                Log.error("There is a problem starting to mine in this pool", Log.LOG_TYPE.POOLS, exception);
-
-            }
-
-        }, undefined, undefined, false);
-
-    }
-
-    async createMiningPool() {
-
-        Log.info('Create Mining Pool', Log.LOG_TYPE.info);
-
-        await this._callCallbackBlockchainSync(async () => {
-
-            try {
-
-                await Blockchain.PoolManagement.setPoolStarted(false);
-
-                let getNewLink = true;
-
-                Log.warn('To be accessible by Browser miners you need an authorized SSL certificate and a free domain.', Log.LOG_TYPE.info);
-
-                if (typeof Blockchain.PoolManagement.poolSettings.poolURL === "string" && Blockchain.PoolManagement.poolSettings.poolURL !== '') {
-
-                    console.info('You have some settings for a pool: ', Blockchain.PoolManagement.poolSettings.poolName, " ", Blockchain.PoolManagement.poolSettings.poolWebsite);
-                    let response = await AdvancedMessages.confirm('Do you want to continue using the settings for : ' + Blockchain.PoolManagement.poolSettings.poolURL);
-
-                    if (response === true) getNewLink = false;
-                }
-
-                if (getNewLink) {
-
-                    let poolFee, poolReferralFee, poolName, poolWebsite, poolServers;
-
-
-                    poolFee = await AdvancedMessages.readNumber('Choose a fee(0...100): ', true);
-
-                    if (isNaN(poolFee) || poolFee < 0 || 100 < poolFee) {
-                        Log.error("You have entered an invalid number: " + poolFee, Log.LOG_TYPE.POOLS);
-                        return false;
-                    } else
-                        Log.info("Your fee is " + poolFee, Log.LOG_TYPE.POOLS);
-
-                    poolName = await AdvancedMessages.input('Pool Name: ');
-                    poolWebsite = await AdvancedMessages.input('Pool Website: ');
-
-                    poolReferralFee = await AdvancedMessages.readNumber("Choose a Referral fee (0...100): ", true);
-                    if (isNaN(poolReferralFee) || poolReferralFee < 0 || 100 < poolReferralFee) {
-                        Log.error("You have entered an invalid number:" + poolReferralFee, Log.LOG_TYPE.POOLS);
-                        return false;
-                    } else
-                        Log.warn("Your Referral fee is: " + poolReferralFee, Log.LOG_TYPE.POOLS);
-
-                    let response = await AdvancedMessages.confirm('Do you want to use external pool servers?: ');
-
-                    if (response)
-                        poolServers = await AdvancedMessages.input('Pool Servers (separated by comma): ');
-                    else
-                        poolServers = await NodeServer.getServerHTTPAddress(true);
-
-                    console.info("Pool Servers:", poolServers);
-
-                    if (response) {
-                        await Blockchain.PoolManagement.poolSettings.setPoolUsePoolServers(true);
-                    } else {
-                        await Blockchain.PoolManagement.poolSettings.setPoolUsePoolServers(false);
-                        await Blockchain.PoolManagement.poolSettings.setPoolUseSignatures(false);
-                    }
-
-
-                    if (poolFee) await Blockchain.PoolManagement.poolSettings.setPoolFee(poolFee / 100);
-                    if (poolName) await Blockchain.PoolManagement.poolSettings.setPoolName(poolName);
-                    if (poolWebsite) await Blockchain.PoolManagement.poolSettings.setPoolWebsite(poolWebsite);
-                    if (poolServers) await Blockchain.PoolManagement.poolSettings.setPoolServers(poolServers);
-                    if (poolReferralFee) await Blockchain.PoolManagement.poolSettings.setPoolReferralFee(poolReferralFee / 100);
-
-                }
-
-            } catch (exception) {
-
-                Log.error("Error starting your pool", Log.LOG_TYPE.POOLS, exception);
-
-            }
-
-        }, async () => {
-
-            await Blockchain.PoolManagement.startPool(true);
-
-        }, undefined, undefined, true);
-
-    }
-
-    async createServerForMiningPool() {
-
-        await this._callCallbackBlockchainSync(async () => {
-
-                console.info('Create Server Pool');
-                console.warn('To be accessible by Browser miners you need an authorized SSL certificate and a free domain.');
-
-                let serverPoolFee = await AdvancedMessages.readNumber('Choose a fee(0...100): ', true);
-
-                if (isNaN(serverPoolFee) || serverPoolFee < 0 || 100 < serverPoolFee) {
-                    console.log("You have entered an invalid number:", serverPoolFee);
-                    return false;
-                } else
-                    console.log("your fee is", serverPoolFee);
-
-                await Blockchain.ServerPoolManagement.serverPoolSettings.setServerPoolFee(serverPoolFee / 100);
-
-            },
-            async () => {
-
-                await Blockchain.ServerPoolManagement.startServerPool();
-
-            }, undefined, undefined, true,
-        );
-
-
-    }
-
-    async _callCallbackBlockchainSync(callbackBeforeBlockchainLoaded, callbackBeforeServerInitialization, callbackAfterServerInitialization, afterSynchronizationCallback, synchronize = true) {
-
-        if (!Blockchain._blockchainInitiated) {
-
-            await Blockchain.createBlockchain("full-node", callbackBeforeBlockchainLoaded, async () => {
-
-                if (callbackBeforeServerInitialization) await callbackBeforeServerInitialization();
-
-                await Node.NodeServer.startServer();
-
-                await Node.NodeClientsService.startService();
-
-                if (callbackAfterServerInitialization) await callbackAfterServerInitialization();
-
-            }, afterSynchronizationCallback, synchronize);
-
-        } else {
-
-            if (callbackBeforeBlockchainLoaded) await callbackBeforeBlockchainLoaded();
-
-            if (callbackBeforeServerInitialization) await callbackBeforeServerInitialization();
-
-            if (callbackAfterServerInitialization) await callbackAfterServerInitialization();
-
-            if (afterSynchronizationCallback) await afterSynchronizationCallback();
-
-        }
-
-    }
-
-
-    disableForksImmutability() {
-
-        consts.BLOCKCHAIN.FORKS.IMMUTABILITY_LENGTH += 10000;
-
-        setTimeout(() => {
-
-            consts.BLOCKCHAIN.FORKS.IMMUTABILITY_LENGTH -= 10000;
-
-        }, 10 * 60 * 1000);
-
-    }
-
-    disconnectFromAllConnectedNodes() {
-
-        let NodesList = require('node/lists/Nodes-List').default;
-        NodesList.disconnectAllNodes(CONNECTION_TYPE.CONNECTION_CLIENT_SOCKET);
-
-    }
-
-    disconnectAllMinersNodes() {
-
-        let NodesList = require('node/lists/Nodes-List').default;
-        NodesList.disconnectAllNodesByConsensusType(NODES_CONSENSUS_TYPE.NODE_CONSENSUS_MINER_POOL);
-
-    }
-
-    async setIntervalDisconnectAllMinersNodes() {
-
-        let intervalTime = await AdvancedMessages.readNumber('Enter the interval time: \n 0 - disable interval \n x - minutes');
-
-        if (this._intervalDisconnectingMiners)
-            clearTimeout(this._intervalDisconnectingMiners);
-
-        if (intervalTime > 0) {
-
-            this._intervalDisconnectingMiners = setInterval(() => {
-                this.disconnectAllMinersNodes();
-            }, intervalTime * 60 * 1000);
-
-        }
-
-    }
-
 }
-
-const commands = [
-    '1. List addresses',
-    '2. Create new address',
-    '3. Delete address',
-    '4. Import address',
-    '5. Export address',
-    '6. Encrypt address',
-    '7. Set mining address',
-    '8. Solo: Start Mining',
-    '9. Solo: Start Mining Instantly Even Unsynchronized',
-    '10. Mining Pool: Start Mining',
-    '11. Mining Pool: Create a New Pool',
-    '11-1. Mining Pool: Process Remaining Payment',
-    '12. Server for Mining Pool: Create a new Server for Mining Pool (Optional and Advanced)',
-    '13. Create Offline Transaction',
-    '20. HTTPS Express Start',
-    '21. Disable Node Immutability',
-    '22. Disconnect from all consensus nodes',
-    '23. Disconnect all miner nodes',
-    '24. Set Interval to disconnect all miner nodes',
-    '30. Set Password for Mining Address',
-    '53. Add address to banlist',
-    '54. Load address banlist',
-    '55. Save address banlist',
-];
-
 
 const lineSeparator =
     "\n|_______|____________________________________________|_________________|";
